Rework diagnostics hud
