+++ src/main/java/org/orecruncher/dsurround/Client.java
@@ -10,6 +10,7 @@ import org.orecruncher.dsurround.config.*;
 import org.orecruncher.dsurround.config.libraries.*;
 import org.orecruncher.dsurround.config.libraries.impl.*;
 import org.orecruncher.dsurround.effects.particles.ParticleSheets;
+import org.orecruncher.dsurround.gui.hud.DiagnosticsOverlay;
 import org.orecruncher.dsurround.gui.hud.OverlayManager;
 import org.orecruncher.dsurround.gui.keyboard.KeyBindings;
 import org.orecruncher.dsurround.lib.GameUtils;
@@ -25,7 +26,6 @@ import org.orecruncher.dsurround.lib.version.VersionResult;
 import org.orecruncher.dsurround.processing.Handlers;
 import org.orecruncher.dsurround.runtime.ConditionEvaluator;
 import org.orecruncher.dsurround.runtime.IConditionEvaluator;
-import org.orecruncher.dsurround.runtime.diagnostics.*;
 import org.orecruncher.dsurround.sound.IAudioPlayer;
 import org.orecruncher.dsurround.sound.MinecraftAudioPlayer;
 
@@ -80,7 +80,6 @@ public class Client implements IMinecraftMod, ClientModInitializer {
         ContainerManager.getDefaultContainer()
                 .registerSingleton(Config)
                 .registerSingleton(IConditionEvaluator.class, ConditionEvaluator.class)
-                .registerSingleton(Diagnostics.class)
                 .registerSingleton(IVersionChecker.class, VersionChecker.class)
                 .registerSingleton(ISoundLibrary.class, SoundLibrary.class)
                 .registerSingleton(IBiomeLibrary.class, BiomeLibrary.class)
@@ -89,6 +88,7 @@ public class Client implements IMinecraftMod, ClientModInitializer {
                 .registerSingleton(IItemLibrary.class, ItemLibrary.class)
                 .registerSingleton(IEntityEffectLibrary.class, EntityEffectLibrary.class)
                 .registerSingleton(IAudioPlayer.class, MinecraftAudioPlayer.class)
+                .registerSingleton(DiagnosticsOverlay.class)
                 .registerSingleton(OverlayManager.class);
 
         // Kick off version checking if configured.  This should run in parallel with initialization.
