+++ src/main/java/org/orecruncher/dsurround/gui/hud/plugins/SoundEngineDiagnosticsPlugin.java
@@ -1,4 +1,4 @@
-package org.orecruncher.dsurround.runtime.diagnostics;
+package org.orecruncher.dsurround.gui.hud.plugins;
 
 import joptsimple.internal.Strings;
 import net.fabricmc.api.EnvType;
@@ -7,6 +7,7 @@ import net.minecraft.client.sound.Channel;
 import net.minecraft.client.sound.SoundInstance;
 import net.minecraft.util.Formatting;
 import org.orecruncher.dsurround.eventing.ClientEventHooks;
+import org.orecruncher.dsurround.gui.hud.IDiagnosticPlugin;
 import org.orecruncher.dsurround.lib.GameUtils;
 import org.orecruncher.dsurround.lib.events.HandlerPriority;
 import org.orecruncher.dsurround.mixins.audio.MixinSoundManagerAccessor;
@@ -17,11 +18,11 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 
 @Environment(EnvType.CLIENT)
-public class SoundEngineDiagnostics implements IDiagnosticPlugin {
+public class SoundEngineDiagnosticsPlugin implements IDiagnosticPlugin {
 
     private static final String FMT_DBG_SOUND = Formatting.GOLD + "%s: %d";
 
-    public SoundEngineDiagnostics() {
+    public SoundEngineDiagnosticsPlugin() {
         ClientEventHooks.COLLECT_DIAGNOSTICS.register(this::onCollect, HandlerPriority.LOW);
     }
