+++ src/main/java/org/orecruncher/dsurround/gui/keyboard/KeyBindings.java
@@ -5,11 +5,11 @@ import net.fabricmc.api.Environment;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 import net.minecraft.client.option.KeyBinding;
 import net.minecraft.client.util.InputUtil;
+import org.orecruncher.dsurround.gui.hud.DiagnosticsOverlay;
 import org.orecruncher.dsurround.gui.sound.IndividualSoundControlScreen;
 import org.orecruncher.dsurround.lib.GameUtils;
 import org.orecruncher.dsurround.lib.di.ContainerManager;
 import org.orecruncher.dsurround.lib.infra.events.ClientState;
-import org.orecruncher.dsurround.runtime.diagnostics.Diagnostics;
 import org.orecruncher.dsurround.sound.IAudioPlayer;
 
 @Environment(EnvType.CLIENT)
@@ -49,7 +49,7 @@ public class KeyBindings {
         ClientState.TICK_END.register(client -> {
             if (GameUtils.getCurrentScreen() == null && GameUtils.getPlayer() != null) {
                 if (diagnosticHud.wasPressed())
-                    ContainerManager.resolve(Diagnostics.class).toggleCollection();
+                    ContainerManager.resolve(DiagnosticsOverlay.class).toggleCollection();
             }
         });
     }
